add cross compilation of linux or darwin
