+++ audio-streaming-client-go/baiduasr/README.md
@@ -1,0 +1,15 @@
+# 运行
+```shell script
+# 可通过 go run main.go -h 查看可输入的命令
+go run main.go -server_addr=127.0.0.1:8051 -username 123 -password 123 
+
+# 编译二进制
+GOOS=darwin GOARCH=arm64 go build
+```
+
+# 注
+mac上运行之前，需要执行
+```shell
+brew install pkg-config
+brew install portaudio
+```
